run build_deps only if built is not present

+++ helpers/deps/__init__.py
@@ -34,14 +34,15 @@ def clean_deps():
     os.chdir(prefix)
 
 def build_deps(configure):
-    prefix=os.getcwd()
-    os.chdir('extern')
-    os.chmod(configure[0], 0755)
-    out = call(configure)
-    if out != 0: exit(out)
-#    cflags = '-fPIC'
-#    out = call(['make', 'CFLAGS='+cflags,'-j'+str(cpu_count()+1), 'install'])
-    out = call(['make', '-j'+str(cpu_count()+1), 'install'])
-    if out != 0: exit(out)
-    open('built', 'w').close()
-    os.chdir(prefix)
+    if not os.path.exists('extern/built'):
+        prefix=os.getcwd()
+        os.chdir('extern')
+        os.chmod(configure[0], 0755)
+        out = call(configure)
+        if out != 0: exit(out)
+    #    cflags = '-fPIC'
+    #    out = call(['make', 'CFLAGS='+cflags,'-j'+str(cpu_count()+1), 'install'])
+        out = call(['make', '-j'+str(cpu_count()+1), 'install'])
+        if out != 0: exit(out)
+        open('built', 'w').close()
+        os.chdir(prefix)
